requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
pytest
pylint
black
python-multipart
openai
sqlalchemy
pgvector
pypdf
sqlalchemy 
psycopg2-binary
phi
psycopg[binary]
streamlit
google-generativeai
yt-dlp
